split intermediate planning into multiple files
