+++ crates/rayexec_execution/src/execution/intermediate/planner/plan_project.rs
@@ -1,0 +1,42 @@
+use std::sync::Arc;
+
+use rayexec_error::{Result, ResultExt};
+
+use super::{IntermediatePipelineBuildState, Materializations, PipelineIdGen};
+use crate::execution::intermediate::pipeline::IntermediateOperator;
+use crate::execution::operators::project::ProjectOperation;
+use crate::execution::operators::simple::SimpleOperator;
+use crate::execution::operators::PhysicalOperator;
+use crate::logical::logical_project::LogicalProject;
+use crate::logical::operator::{LogicalNode, Node};
+
+impl<'a> IntermediatePipelineBuildState<'a> {
+    pub fn plan_project(
+        &mut self,
+        id_gen: &mut PipelineIdGen,
+        materializations: &mut Materializations,
+        mut project: Node<LogicalProject>,
+    ) -> Result<()> {
+        let location = project.location;
+
+        let input = project.take_one_child_exact()?;
+        let input_refs = input.get_output_table_refs(self.bind_context);
+        self.walk(materializations, id_gen, input)?;
+
+        let projections = self
+            .expr_planner
+            .plan_scalars(&input_refs, &project.node.projections)
+            .context("Failed to plan expressions for projection")?;
+
+        let operator = IntermediateOperator {
+            operator: Arc::new(PhysicalOperator::Project(SimpleOperator::new(
+                ProjectOperation::new(projections),
+            ))),
+            partitioning_requirement: None,
+        };
+
+        self.push_intermediate_operator(operator, location, id_gen)?;
+
+        Ok(())
+    }
+}
